pyproject.toml
[tool.poetry]
name = "repl_python3_DiscordBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
py-cpuinfo = "^8.0.0"
requests-html = "^0.10.0"
lxml = "^4.6.4"
cchardet = "^2.1.7"
psutil = "^5.8.0"
bs4 = "^0.0.1"
discord = "^1.7.3"
youtube-search-python = "^1.4.9"
PyNaCl = "^1.4.0"
yt-dlp = "^2021.12.1"
pytz = "^2021.3"
replit = "^3.2.4"
cryptography = "^36.0.1"
genshin = "^0.3.1"
AnilistPython = "^0.1.3"
youtube_dl = "^2021.12.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
